+++ src/test/java/com/ggukmoney/beanzip/support/RedisIntegrationTestSupport.java
@@ -40,7 +40,7 @@ public abstract class RedisIntegrationTestSupport {
         redisTemplate = new StringRedisTemplate(connectionFactory);
         redisTemplate.afterPropertiesSet();
         RedisService redisService = new RedisService(redisTemplate);
-        authService = new AuthService(null, redisService, null, null, null, null, null, null, null);
+        authService = new AuthService(null, redisService, null, null, null, null);
         tapBatchService = new TapBatchService(null, null, null, null, null, null, null, redisService, null, null);
         flushRedis();
     }